Crate wcstr

Rust FFI helpers for working with win32 API's "Unicode" functions that uses "wide" strings.

Structs

NoNulError
An error returned when an expected nul is not found in the string, slice or vector provided.

NulError
An error returned when an unexpected nul is found in the string, slice or vector provided.

Split
Created with method .split(delim)

WCStr
Representation of a borrowed Win32 style "wide" string.

WCString
A type representing an owned Win32 style "wide" string.