Struct wcstr::WCStr

pub struct WCStr {
    // some fields omitted
}

Representation of a borrowed Win32 style "wide" string.

Methods

impl WCStr

unsafe fn from_raw_parts<'a>(ptr: *const u16, len: usize) -> &'a WCStr

Create a &WCStr from a raw pointer and a length.

This function is unsafe for the reasons mentioned below.

This function assumes that the pointer passed in has these properties:

It is not null.

It is a valid pointer.

It points to an array of u16's that does not contain any nul values.

It points to an array of u16's that is terminated with a nul at exactly the offset "len".

This function will assert/panic when nul is not found at offset "len".

The lifetime of the &WCStr returned from this function is not guranteed to be correct and it is up to the caller to determine the appropriate lifetime.

from_slice_with_nul() should be used instead of this function when possible.

from_raw_parts() example

use wcstr::WCStr;
static a : &'static [u16] = &[116u16, 101u16, 115u16, 116u16, 0];
let s = unsafe { WCStr::from_raw_parts(a.as_ptr(), a.len() - 1) };
assert!(s.len() == (a.len() - 1));

fn from_slice_with_nul<'a>(slice: &'a [u16]) -> Result<&'a WCStr, NoNulError>

Create a &WCStr from a slice of u16's. This function will scan the slice for nul and assume that nul terminates the string. If no nul is found in the slice, it will return Err(NoNulError(None))

frm_slice_with_nul() example

use wcstr::WCStr;
static a : &'static [u16] = &[116u16, 101u16, 115u16, 116u16, 0];
let s = WCStr::from_slice_with_nul(a).unwrap();
assert!(s.len() == (a.len() - 1));

fn len(&self) -> usize

length of the string in u16 units

fn is_empty(&self) -> bool

is empty

fn as_ptr(&self) -> *const u16

Return a raw pointer to this "wide" string.

The pointer remains valid only as long as this string is valid.

The pointer points to a contiguous region of memory terminated with nul.

fn to_slice(&self) -> &[u16]

Return this "wide" string as a slice of u16s without a nul terminator.

fn to_slice_with_nul(&self) -> &[u16]

Return this "wide" string as a slice of u16s with a nul terminator.

fn to_string(&self) -> Result<String, FromUtf16Error>

Convert this "wide" string to a String by using String::from_utf16

fn to_string_lossy(&self) -> String

Convert this "wide" string to a String by using String::from_utf16_lossy

fn to_os_string(&self) -> OsString

Convert this "wide" string to an OsString by using OsString::from_wide

fn to_path_buf(&self) -> PathBuf

Convert this "wide" string to a PathBuf

fn starts_with<T>(&self, s: T) -> bool where T: AsRef<WCStr>

starts with a string.

starts_with() example

use wcstr::{WCStr, WCString};
let s = WCString::from_str("abcefg").unwrap();
let t = WCString::from_str("abc").unwrap();
let u = WCString::from_str("efg").unwrap();
let v = WCString::from_str("abcefgh").unwrap();
assert!(s.starts_with(t));
assert!(!s.starts_with(u));
assert!(!s.starts_with(v));

fn starts_with_str<T>(&self, s: T) -> bool where T: AsRef<OsStr>

starts with a string.

starts_with_str() example

use wcstr::{WCStr, WCString};
let s = WCString::from_str("abcefg").unwrap();
assert!(s.starts_with_str("abc"));
assert!(!s.starts_with_str("efg"));
assert!(!s.starts_with_str("abcefgh"));