Struct wcstr::WCString

pub struct WCString {
    // some fields omitted
}

A type representing an owned Win32 style "wide" string.

Methods

impl WCString

fn new() -> WCString

Create an empty WCString.

new() example

use wcstr::WCString;
let s = WCString::new();
assert!(s.len() == 0);

fn from_vec<T>(v: T) -> Result<WCString, NulError> where T: Into<Vec<u16>>

Create a WCString from a Vec<u16>. The string will be scanned for nul and NulError will be returned if a nul is found.

from_vec() example

use wcstr::WCString;
use std::os::windows::ffi::OsStrExt;
use std::ffi::OsStr;
let v: Vec<_> = OsStr::new("testing").encode_wide().collect();
let s = WCString::from_vec(v).unwrap();
assert!(s.len() == 7);

fn from_vec_with_nul<T>(u16s: T) -> Result<WCString, NoNulError> where T: Into<Vec<u16>>

Create a WCString from a Vec<u16> with a nul terminator. The string will be scanned for nul. The string will be truncated at the position where nul is found. NoNulError will be returned if a nul could not be found.

from_vec_with_nul() example

use wcstr::WCString;
use std::os::windows::ffi::OsStrExt;
use std::ffi::OsStr;
let v: Vec<_> = OsStr::new("testing\0").encode_wide().collect();
let s = WCString::from_vec_with_nul(v).unwrap();
assert!(s.len() == 7);

unsafe fn from_vec_unchecked(v: Vec<u16>) -> WCString

Create a WCString from a Vec<u16> without checking for validity. This function is unsafe as it assumes that the string passed in has no nul in it.

from_vec_unchecked() example

use wcstr::WCString;
use std::os::windows::ffi::OsStrExt;
use std::ffi::OsStr;
let v: Vec<_> = OsStr::new("testing").encode_wide().collect();
let s = unsafe { WCString::from_vec_unchecked(v) };
assert!(s.len() == 7);

unsafe fn from_vec_with_nul_unchecked(v: Vec<u16>) -> WCString

Create a WCString from a Vec<u16> with a nul terminator without checking for validity. This function is unsafe for the following reasons: * This function assumes that the string passed in has no nul in it aside from the nul terminator. * This function assumes that the string passed in has a nul terminator at the end.

from_vec_with_nul_unchecked() example

use wcstr::WCString;
use std::os::windows::ffi::OsStrExt;
use std::ffi::OsStr;
let v: Vec<_> = OsStr::new("testing\0").encode_wide().collect();
let s = unsafe { WCString::from_vec_with_nul_unchecked(v) };
assert!(s.len() == 7);

fn from_str<T>(s: T) -> Result<WCString, NulError> where T: AsRef<OsStr>

Create a WCString from a &OsStr (or anything that can be cast to &OsStr, including OsString, &str,&Path,PathBufandString) The string will be scanned fornulandNulErrorwill be returned if anul` is found.

from_str() example

use wcstr::WCString;
let s = WCString::from_str("testing").unwrap();
assert!(s.len() == 7);

fn from_str_with_nul<T>(s: T) -> Result<WCString, NoNulError> where T: AsRef<OsStr>

Create a WCString from a &OsStr (or anything that can be cast to &OsStr``) with anulterminator. The string will be scanned fornulandNoNulErrorwill be returned if anulcould not be found. The string will be truncated at the position wherenul`` is found.

from_str_with_nul() example

use wcstr::WCString;
let s = WCString::from_str_with_nul("testing\0ignored").unwrap();
assert!(s.len() == 7);

fn into_vec(self) -> Vec<u16>

Return the underlying buffer as a Vec<u16>.

The WCString will be consumed.

The returned buffer does not contain the nul terminator.

The returned buffer does not contain any nul. # into_vec() example use wcstr::WCString; let s = WCString::from_str("testing").unwrap(); let v = s.into_vec(); assert!(*v.last().unwrap() != 0);

fn into_vec_with_nul(self) -> Vec<u16>

Return the underlying buffer as a Vec<u16> with a nul terminator. The WCString will be consumed. The returned buffer does not contain any nul aside from the nul terminator.

into_vec_with_nul() example

use wcstr::WCString;
let s = WCString::from_str("testing").unwrap();
let v = s.into_vec_with_nul();
assert!(*v.last().unwrap() == 0);

fn as_slice(&self) -> &[u16]

Return the underlying buffer as a u16 slice.

The returned slice does not contain the nul terminator.

The returned slice does not contain any nul. # as_slice() example use wcstr::WCString; let s = WCString::from_str("testing").unwrap(); let w = s.as_slice(); assert!(*w.last().unwrap() != 0);

fn as_slice_with_nul(&self) -> &[u16]

Return the underlying buffer as a u16 slice with a nul terminator. The returned slice does not contain any nul aside from the nul terminator.

as_slice_with_nul() example

use wcstr::WCString;
let s = WCString::from_str("testing").unwrap();
let w = s.as_slice_with_nul();
assert!(*w.last().unwrap() == 0);

fn as_wcstr(&self) -> &WCStr

Return this string as a &WCStr

as_wcstr() example

use wcstr::WCString;
let s = WCString::from_str("testing").unwrap();
let w = s.as_wcstr();

fn push<T>(&mut self, s: T) where T: AsRef<WCStr>

Push/Append a &WCStr (or anything that can cast to a &WCStr, like another WCString).

push() example

use wcstr::WCString;
let mut s = WCString::new();
let t = WCString::from_str("test").unwrap();
s.push(&t);
s.push(&t);

fn push_slice<T>(&mut self, s: T) -> Result<(), NulError> where T: AsRef<[u16]>

Push/Append a u16 slice. The slice will be scanned for nul, and the push will fail with NulError if a nul is found.

push_slice() example

use wcstr::WCString;
let mut s = WCString::new();
let t = WCString::from_str("test").unwrap();
let t = t.as_slice();
s.push_slice(t).unwrap();
s.push_slice(t).unwrap();

fn push_slice_with_nul<T>(&mut self, s: T) -> Result<(), NoNulError> where T: AsRef<[u16]>

Push/Append a u16 slice with a nul terminator. The slice will be scanned for nul, and the push will fail with NoNulError if a nul is not found. The push will stop at the first nul found in the slice, and the remaining part of the slice will be ignored.

push_slice_with_nul() example

use wcstr::WCString;
let mut s = WCString::new();
let t = WCString::from_str("test").unwrap().into_vec_with_nul();
s.push_slice_with_nul(&t).unwrap();
s.push_slice_with_nul(&t).unwrap();

fn push_str<T>(&mut self, s: T) -> Result<(), NulError> where T: AsRef<OsStr>

Push/Append a &OsStr (or anything that can be cast to &OsStr) The string will be scanned for nul, and the push will fail with NulError if a nul is found.

push_str() example

use wcstr::WCString;
let mut s = WCString::new();
s.push_str("test1").unwrap();
s.push_str("test2").unwrap();

fn push_str_with_nul<T>(&mut self, s: T) -> Result<(), NoNulError> where T: AsRef<OsStr>

Push/Append a &OsStr (or anything that can be cast to &OsStr) The string will be scanned for nul, and the push will fail with NoNulError if a nul is not found. The push will stop at the first nul found in the string.

push_str_with_nul() example

use wcstr::WCString;
let mut s = WCString::new();
s.push_str_with_nul("test1\0everything after nul will be ignored").unwrap();
s.push_str_with_nul("test2\0").unwrap();

fn split(self, delimiter: u16) -> Split

Split the string into multiple &mut WCStr using a delimiter.

This returns an iterator that creates a &mut WCStr for each part of the string separated by the delimiter.

This will consume the string.

split() example

use wcstr::WCString;
let s = WCString::from_str("a;b;c;d;e").unwrap();
let mut count = 0;
for w in s.split(b';' as u16).iter() {
    count += 1;
    assert!(w.len() == 1);
}
assert!(count == 5);

fn replace(&mut self, needle: u16, replacement: u16)

Replace a u16 value with another u16 value in the string.

This will assert if either needle or replacement is nul

replace() example

use wcstr::WCString;
let mut s = WCString::from_str("x y z").unwrap();
s.replace(b' ' as u16, b':' as u16);
let t = s.to_string().unwrap();
assert!(t == "x:y:z");